Prevent outdated static files
Added a hashcode to the file names of the css files. That way, when the
content changes, the browser will always fetch the new file, instead of
using the locally cached css that may be outdated.
This could also be applied to js files (there are none atm), but not
to images, as it would require search and replaces in both html and
css, while image changes are not so likely.